Remove byte-pool dependency because the interface does not work for our use-case.
A type that can be used as backing implementation for the byte pool needs to implement Poolable. But the interface contains the function

fn alloc(size: usize) -> Self

which we can't implement, since it does allow any state as argument. But we need to have a reference to the actual memory mapped file and can't create new containers without state from memory.