Move BtreeIndex implementation to module outside_keys.
This implementation is generic for all types of keys that can be serialized with serde.
Thus it needs a separate tuple storage to store the actual key value. We also want
to add an implementation where the key is stored inline, because the key small enough an can be serialized without any help of serde.